+++ src/lib/src/components/default-error.component.ts
@@ -1,4 +1,5 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { ActivatedView } from '../activated-view';
 import { ViewData } from '../view-data';
 
 @Component({
@@ -6,8 +7,14 @@ import { ViewData } from '../view-data';
   <p>{{data.status}} {{data.statusText}}</p>
   <pre *ngIf="data.body">{{data.body}}</pre>`
 })
-export class DefaultErrorComponent {
+export class DefaultErrorComponent implements OnInit {
 
-  constructor(public data: ViewData<any>) {
+  public data: ViewData<any>;
+
+  constructor(public view: ActivatedView<any>) {
+  }
+
+  ngOnInit(): void {
+    this.view.data.subscribe(data => this.data = data);
   }
 }