+++ src/lib/src/directives/resource-data.ts
@@ -3,7 +3,7 @@ import 'rxjs/add/operator/switchMap';
 import { Observable } from 'rxjs/Observable';
 import { ViewDataLoader } from '../view-data-loader';
 import { ViewData } from '../view-data';
-import { Navigable } from '../navigation-handler';
+import { Navigable } from '../navigable';
 import { ResourceViewRegistry } from '../resource-view-registry';
 import { MEDIA_TYPE_ROUTER_EMPTY, MEDIA_TYPE_ROUTER_ERROR, MEDIA_TYPE_ROUTER_LOADING } from '../system-media-types';
 import { Subject } from 'rxjs/Subject';
@@ -65,7 +65,8 @@ export class ResourceDataDirective implements OnInit, Navigable {
   }
 
   set url(value: string) {
-    // This is needed check, since during runtime binding, we cannot be sure value isn't null
+    // Normalize undefined (from binding) to empty string
+    // TODO this might break binding change detection
     if (!value) {
       value = '';
     }